[YS-531] refactor: 바텀 시트 위에 모달이 뜨도록 처리
